Fix models

+++ db/migrate/20171017141143_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[5.1]
     create_table :reviews do |t|
       t.belongs_to :repo, null: false
       t.belongs_to :user, null: false
-      t.integer :rating, null: false, default: nil
+      t.float :rating, null: false, default: nil
       t.text :body, null: false
       t.integer :upvotes, null: false, default: 0
       t.integer :downvotes, null: false, default: 0
